requirements.txt
# ─── Servidor web ──────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1

# ─── API de Claude ─────────────────────────────────────────────
anthropic==0.39.0

# ─── Configuración ─────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Testing ───────────────────────────────────────────────────
pytest==8.3.3
pytest-cov==5.0.0
